Split the navigation and form code into two separate IIFEs
Eventually, I'll need to make updates to this site so I want the code to be
namespaced and organized. I don't want the navigation code to be intermingled
with the contact form code, but I don't think there's enough code to warrant
splitting them into different files at this point.