Built theme toggle
I built the markup, styling, and functionality for a
theme toggle. The user's preference will be stored in
local storage and loaded whenever they open the page. If
it's the user's first time visiting the site, I use their
system preference to determine the default theme and used
light as a fallback