Fixed theme flickering
If dark mode was enabled, the page's theme would flicker upon load. The theme
toggle was rendering before the theme classes were applied, so the theme
swap was visible. To fix this, I moved some of the theme logic into inline
script tags, so it would load as fast as possible. Now, there is no visible
flicker when the page is loaded in dark mode.